adadddd

+++ ECC/ClassifierChain.cpp
@@ -1,6 +1,6 @@
 #include "ClassifierChain.hpp"
 
-ClassifierChain::ClassifierChain(int _numValues, std::vector<int> _orderedLabels, int _maxLevel, int _forestSize)
+ClassifierChain::ClassifierChain(size_t _numValues, std::vector<int> _orderedLabels, size_t _maxLevel, size_t _forestSize)
 	: maxLevel(_maxLevel), forestSize(_forestSize), chainSize(_orderedLabels.size()), orderedLabels(_orderedLabels),
 	forests(), treeSize(pow(2, maxLevel + 1) - 1), totalSize(forestSize * treeSize * chainSize)
 {
@@ -13,7 +13,7 @@ ClassifierChain::ClassifierChain(int _numValues, std::vector<int> _orderedLabels
 	}
 }
 
-ClassifierChain::ClassifierChain(int numValues, int numLabels, int maxLevel, int forestSize)
+ClassifierChain::ClassifierChain(size_t numValues, size_t numLabels, size_t maxLevel, size_t forestSize)
 	: ClassifierChain(numValues, standardLabelOrder(numLabels), maxLevel, forestSize)
 {
 }
@@ -28,40 +28,40 @@ const std::vector<int> ClassifierChain::getLabelOrder() const
 	return orderedLabels;
 }
 
-int ClassifierChain::getMaxLevel()
+size_t ClassifierChain::getMaxLevel()
 {
 	return maxLevel;
 }
 
-int ClassifierChain::getTreeSize()
+size_t ClassifierChain::getTreeSize()
 {
 	return treeSize;
 }
 
-int ClassifierChain::getForestSize()
+size_t ClassifierChain::getForestSize()
 {
 	return forestSize;
 }
 
-int ClassifierChain::getChainSize()
+size_t ClassifierChain::getChainSize()
 {
 	return chainSize;
 }
 
-int ClassifierChain::getTotalSize()
+size_t ClassifierChain::getTotalSize()
 {
 	return totalSize;
 }
 
-int ClassifierChain::size()
+size_t ClassifierChain::size()
 {
 	return forestSize*treeSize*chainSize;
 }
 
-std::vector<int> ClassifierChain::standardLabelOrder(int numLabels)
+std::vector<int> ClassifierChain::standardLabelOrder(size_t numLabels)
 {
 	std::vector<int> labels(numLabels);
-	for (int n = 0; n < numLabels; n++)
+	for (size_t n = 0; n < numLabels; n++)
 	{
 		labels[n] = n;
 	}
