+++ ECC/ClassifierChain.hpp
@@ -6,27 +6,27 @@
 class ClassifierChain
 {
 private:
-	const int maxLevel;
-	const int treeSize;
-	const int forestSize;
-	const int chainSize;
-	const int totalSize;
+	const size_t maxLevel;
+	const size_t treeSize;
+	const size_t forestSize;
+	const size_t chainSize;
+	const size_t totalSize;
 
 	std::vector<Forest> forests;
     std::vector<int> orderedLabels;
 public:
-	ClassifierChain(int _numValues, std::vector<int> _orderedLabels, int _maxLevel, int _forestSize);
-	ClassifierChain(int numValues, int numLabels, int maxLevel, int forestSize);
+	ClassifierChain(size_t _numValues, std::vector<int> _orderedLabels, size_t _maxLevel, size_t _forestSize);
+	ClassifierChain(size_t numValues, size_t numLabels, size_t maxLevel, size_t forestSize);
 
 	const std::vector<Forest>& getForests();
 	const std::vector<int> getLabelOrder() const;
-	int getMaxLevel();
-	int getTreeSize();
-	int getForestSize();
-	int getChainSize();
-	int getTotalSize();
-	int size();
+	size_t getMaxLevel();
+	size_t getTreeSize();
+	size_t getForestSize();
+	size_t getChainSize();
+	size_t getTotalSize();
+	size_t size();
 
-	static std::vector<int> standardLabelOrder(int numLabels);
+	static std::vector<int> standardLabelOrder(size_t numLabels);
 };
