+++ ECC/Forest.cpp
@@ -1,9 +1,9 @@
 #include "Forest.hpp"
-Forest::Forest(int _numValues, int _maxLevel, int _label, std::vector<int>& _excludeLabelIndices, int _forestSize)
+Forest::Forest(size_t _numValues, size_t _maxLevel, size_t _label, std::vector<int>& _excludeLabelIndices, size_t _forestSize)
 	: maxLevel(_maxLevel), label(_label), excludeLabels(_excludeLabelIndices), forestSize(_forestSize), forest(),
 	treeSize(pow(2, maxLevel + 1) - 1), totalSize(treeSize*forestSize)
 {
-	for (int n = 0; n < forestSize; n++)
+	for (size_t n = 0; n < forestSize; n++)
 	{
 		forest.push_back(Tree(_numValues, maxLevel, label, _excludeLabelIndices));
 	}
@@ -14,32 +14,32 @@ const std::vector<Tree>& Forest::getTrees()
 	return forest;
 }
 
-int Forest::getMaxLevel()
+size_t Forest::getMaxLevel()
 {
 	return maxLevel;
 }
 
-int Forest::getTreeSize()
+size_t Forest::getTreeSize()
 {
 	return treeSize;
 }
 
-int Forest::getForestSize()
+size_t Forest::getForestSize()
 {
 	return forestSize;
 }
 
-int Forest::getTotalSize()
+size_t Forest::getTotalSize()
 {
 	return totalSize;
 }
 
-int Forest::size()
+size_t Forest::size()
 {
 	return forestSize*treeSize;
 }
 
-int Forest::getLabel()
+size_t Forest::getLabel()
 {
 	return label;
 }
