+++ ECC/Tuner.hpp
@@ -8,23 +8,23 @@ class ECCTuner
 {
 private:
 	ECCExecutorNew eccEx;
-	int numChains;
-	int numTrees;
-	int numLabels;
+	size_t numChains;
+	size_t numTrees;
+	size_t numLabels;
 
 	double tuneClassifyStepFunc(atf::configuration config);
 	double tuneClassifyFinalFunc(atf::configuration config);
 	double tuneBuildFunc(atf::configuration config);
 
-	Configuration runBuildTuner(int treesPerRun);
-	Configuration runClassifyStepTuner(int numInstances);
-	Configuration runClassifyFinalTuner(int numInstances);
+	Configuration runBuildTuner(size_t treesPerRun);
+	Configuration runClassifyStepTuner(size_t numInstances);
+	Configuration runClassifyFinalTuner(size_t numInstances);
 public:
-	ECCTuner(int _maxLevel, int _maxAttributes, int _numAttributes, int _numTrees, int _numLabels, int _numChains, int _ensembleSubSetSize, int _forestSubSetSize);
+	ECCTuner(size_t _maxLevel, size_t _maxAttributes, size_t _numAttributes, size_t _numTrees, size_t _numLabels, size_t _numChains, size_t _ensembleSubSetSize, size_t _forestSubSetSize);
 
-	Configuration tuneBuild(ECCData& buildData, int treesPerRun);
-	Configuration tuneClassifyStep(ECCData& buildData, int treesPerRun, ECCData& classifyData, Configuration config);
-	Configuration tuneClassifyFinal(ECCData& buildData, int treesPerRun, ECCData& classifyData, Configuration config);
+	Configuration tuneBuild(ECCData& buildData, size_t treesPerRun);
+	Configuration tuneClassifyStep(ECCData& buildData, size_t treesPerRun, ECCData& classifyData, Configuration config);
+	Configuration tuneClassifyFinal(ECCData& buildData, size_t treesPerRun, ECCData& classifyData, Configuration config);
 };
 
 #endif
